Change package to uk.co.ribot.androidboilerplate
