+++ app/src/main/AndroidManifest.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.hitherejoe.androidboilerplate" >
+    package="uk.co.ribot.androidboilerplate" >
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
-        android:name=".AndroidBoilerplateApplication"
+        android:name="uk.co.ribot.androidboilerplate.AndroidBoilerplateApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".ui.activity.MainActivity"
+            android:name="uk.co.ribot.androidboilerplate.ui.activity.MainActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,10 +21,10 @@
             </intent-filter>
         </activity>
 
-        <service android:name=".data.SyncService"/>
+        <service android:name="uk.co.ribot.androidboilerplate.data.SyncService"/>
 
         <receiver
-            android:name=".data.SyncService$SyncOnConnectionAvailable"
+            android:name="uk.co.ribot.androidboilerplate.data.SyncService$SyncOnConnectionAvailable"
             android:enabled="false">
             <intent-filter>
                 <action android:name="android.net.conn.CONNECTIVITY_CHANGE" />
