--- app/src/main/java/com/hitherejoe/androidboilerplate/AndroidBoilerplateApplication.java
@@ -1,33 +1,0 @@
-package com.hitherejoe.androidboilerplate;
-
-import android.app.Application;
-
-import com.hitherejoe.androidboilerplate.data.DataManager;
-
-import rx.schedulers.Schedulers;
-
-public class AndroidBoilerplateApplication extends Application {
-
-    private static AndroidBoilerplateApplication sAndroidBoilerplateApplication;
-    private DataManager mDataManager;
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-        sAndroidBoilerplateApplication = this;
-        mDataManager = new DataManager(this, Schedulers.io());
-    }
-
-    @Override
-    public void onTerminate() {
-        sAndroidBoilerplateApplication = null;
-        super.onTerminate();
-    }
-
-    public static AndroidBoilerplateApplication get() {
-        return sAndroidBoilerplateApplication;
-    }
-
-    public DataManager getDataManager() { return mDataManager; }
-
-}
