--- app/src/main/java/com/hitherejoe/androidboilerplate/data/local/DbOpenHelper.java
@@ -1,33 +1,0 @@
-package com.hitherejoe.androidboilerplate.data.local;
-
-import android.content.Context;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
-
-public class DbOpenHelper extends SQLiteOpenHelper {
-
-    public static final String DATABASE_NAME = "ribots.db";
-    public static final int DATABASE_VERSION = 1;
-
-    public DbOpenHelper(Context context) {
-        super(context, DATABASE_NAME, null, DATABASE_VERSION);
-    }
-
-    @Override
-    public void onCreate(SQLiteDatabase db) {
-        db.beginTransaction();
-        try {
-            //Uncomment line below if you want to enable foreign keys
-            //db.execSQL("PRAGMA foreign_keys=ON;");
-            db.execSQL(Db.RibotsTable.CREATE);
-            //Add other tables here
-            db.setTransactionSuccessful();
-        } finally {
-            db.endTransaction();
-        }
-    }
-
-    @Override
-    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) { }
-
-}
