--- app/src/main/java/com/hitherejoe/androidboilerplate/util/ComponentUtil.java
@@ -1,33 +1,0 @@
-package com.hitherejoe.androidboilerplate.util;
-
-import android.app.ActivityManager;
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.pm.PackageManager;
-import android.util.Log;
-
-import com.hitherejoe.androidboilerplate.BuildConfig;
-
-public class ComponentUtil {
-
-    public static final String TAG = "ComponentUtil";
-
-    public static void toggleComponent(Context context, Class componentClass, boolean enable) {
-        if (BuildConfig.DEBUG) Log.i(TAG, (enable ? "Enabling " : "Disabling ") + componentClass.getSimpleName());
-        ComponentName componentName = new ComponentName(context, componentClass);
-        PackageManager pm = context.getPackageManager();
-        pm.setComponentEnabledSetting(componentName,
-                enable ? PackageManager.COMPONENT_ENABLED_STATE_ENABLED : PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
-                PackageManager.DONT_KILL_APP);
-    }
-
-    public static boolean isServiceRunning(Context context, Class serviceClass) {
-        ActivityManager manager = (ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE);
-        for (ActivityManager.RunningServiceInfo service : manager.getRunningServices(Integer.MAX_VALUE)) {
-            if (serviceClass.getName().equals(service.service.getClassName())) {
-                return true;
-            }
-        }
-        return false;
-    }
-}
