+++ app/src/main/java/uk/co/ribot/androidboilerplate/data/local/PreferencesHelper.java
@@ -1,0 +1,21 @@
+package uk.co.ribot.androidboilerplate.data.local;
+
+import android.content.Context;
+import android.content.SharedPreferences;
+
+public class PreferencesHelper {
+
+    private static SharedPreferences mPref;
+
+    public static final String PREF_FILE_NAME = "android_boilerplate_pref_file";
+
+
+    public PreferencesHelper(Context context) {
+        mPref = context.getSharedPreferences(PREF_FILE_NAME, Context.MODE_PRIVATE);
+    }
+
+    public void clear() {
+        mPref.edit().clear().apply();
+    }
+
+}
