+++ app/src/main/java/uk/co/ribot/androidboilerplate/util/ViewUtil.java
@@ -1,0 +1,25 @@
+package uk.co.ribot.androidboilerplate.util;
+
+import android.app.Activity;
+import android.content.Context;
+import android.content.res.Resources;
+import android.view.inputmethod.InputMethodManager;
+
+public class ViewUtil {
+
+    public static float pxToDp(float px) {
+        float densityDpi = Resources.getSystem().getDisplayMetrics().densityDpi;
+        return px / (densityDpi / 160f);
+    }
+
+    public static int dpToPx(int dp) {
+        float density = Resources.getSystem().getDisplayMetrics().density;
+        return Math.round(dp * density);
+    }
+
+    public static void hideKeyboard(Activity activity) {
+        InputMethodManager imm = (InputMethodManager) activity.getSystemService(Context.INPUT_METHOD_SERVICE);
+        imm.hideSoftInputFromWindow(activity.getWindow().getDecorView().getWindowToken(), 0);
+    }
+
+}
