--- app/src/test/java/com/hitherejoe/androidboilerplate/util/DefaultConfig.java
@@ -1,6 +1,0 @@
-package com.hitherejoe.androidboilerplate.util;
-
-public class DefaultConfig {
-    //The api level that Roboelectric will use to run the unit tests
-    public static final int EMULATE_SDK = 21;
-}
