+++ app/src/test/java/uk/co/ribot/androidboilerplate/util/DefaultConfig.java
@@ -1,0 +1,6 @@
+package uk.co.ribot.androidboilerplate.util;
+
+public class DefaultConfig {
+    //The api level that Roboelectric will use to run the unit tests
+    public static final int EMULATE_SDK = 21;
+}
